Changed IngesterSettings.MaxContentLengthBytes from long to int because RangeAttribute does not actually support the full range of long values.

+++ DistributedWebCrawler.Core/Configuration/IngesterSettings.cs
@@ -11,8 +11,10 @@ namespace DistributedWebCrawler.Core.Configuration
         [Range(1, 100)]
         public int MaxRedirects { get; init; }
 
-        [Range(1, long.MaxValue)]
-        public long? MaxContentLengthBytes { get; init; }
+        // FIXME: this has to be an int because RangeAttribute does not support the full range of long values
+        // Could implement a custom attribute to allow ContentLength > 2GB.
+        [Range(1, int.MaxValue)]
+        public int? MaxContentLengthBytes { get; init; }
 
         public IEnumerable<string>? IncludeMediaTypes { get; init; }
         public IEnumerable<string>? ExcludeMediaTypes { get; init; }
